requirements.txt
flask>=3.0
firecrawl-py>=4.0
python-dotenv>=1.0
groq>=0.9.0
gunicorn>=21.0
